backend/requirements.txt
langgraph
langchain-core
langchain-google-genai
tavily-python
langchain-community
fastapi
uvicorn
sse-starlette
python-dotenv
pydantic
